check if last school before destroying

+++ app/jobs/sync_contact_schools_to_salesforce_job.rb
@@ -21,8 +21,10 @@ class SyncContactSchoolsToSalesforceJob < ApplicationJob
           contact_id: relation.contact_id,
           school_id: relation.school_id
         )
-        sf_relation&.destroy
-        relation.destroy!
+        if sf_relation.count >= 1
+          sf_relation&.destroy
+          relation.destroy!
+        end
 
       rescue => e
         Sentry.capture_exception e
